Support handling of a KoreSystemObject.
Keep a reference to a foreign object, to be provided to a callback specific to the object. Used specifically for when this library is used through a Foreign Function Interface context, to setup an update loop which calls back to such an object which has state otherwise unaccessible do to FFI callback restrictions.

See pony-kore as an example, which uses KoreC via its FFI.